requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
ultralytics
fastapi
uvicorn[standard]
python-multipart
Pillow
jinja2
